requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.23.10
botocore==1.26.10
Django==3.2.25
django-storages==1.12.3
djangorestframework==3.15.1
drf-yasg==1.21.7
importlib-resources==5.4.0
inflection==0.5.1
jmespath==0.10.0
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.18
zipp==3.6.0
